update electron. use local sha hash and github download (#25155)

+++ shared/desktop/package.desktop.tsx
@@ -8,6 +8,8 @@ import path from 'path'
 import webpack from 'webpack'
 import rootConfig from './webpack.config.babel'
 
+const TEMP_SKIP_BUILD: boolean = false
+
 // absolute path relative to this script
 const desktopPath = (...args: Array<string>) => path.join(__dirname, ...args)
 
@@ -21,26 +23,25 @@ const copySyncFolder = (src: string, target: string, onlyExts: Array<string>) =>
       return !ext || onlyExts.includes(ext)
     },
   })
-  const relSrcs = files.map(f => f.path.substr(srcRoot.length))
+  const relSrcs = files.map(f => f.path.substring(srcRoot.length))
   const dsts = relSrcs.map(f => path.join(dstRoot, f))
 
   relSrcs.forEach((s, idx) => fs.copySync(path.join(srcRoot, s), dsts[idx]))
 }
 
-const copySync = (src, target, options?) => {
+const copySync = (src: string, target: string, options?: object) => {
   fs.copySync(desktopPath(src), desktopPath(target), {...options, dereference: true})
 }
 
-const argv: {[key: string]: any} = minimist(process.argv.slice(2), {string: ['appVersion']})
+const argv = minimist(process.argv.slice(2), {string: ['appVersion']}) as {[key: string]: string | undefined}
 
 const appName = 'Keybase'
-const shouldUseAsar: boolean = argv.asar || argv.a || false
-const shouldBuildAll: boolean = argv.all || false
-const arch: string = argv.arch ? argv.arch.toString() : os.arch()
-const platform: string = argv.platform ? argv.platform.toString() : os.platform()
-const appVersion: string = argv.appVersion || '0.0.0'
-const comment: string = argv.comment || ''
-const outDir: string = argv.outDir || ''
+const shouldUseAsar = false
+const arch: string = typeof argv.arch === 'string' ? argv.arch.toString() : os.arch()
+const platform: string = typeof argv.platform === 'string' ? argv.platform.toString() : os.platform()
+const appVersion: string = (typeof argv.appVersion === 'string' && argv.appVersion) || '0.0.0'
+const comment: string = (typeof argv.comment === 'string' && argv.comment) || ''
+const outDir: string = (typeof argv.outDir === 'string' && argv.outDir) || ''
 const appCopyright = 'Copyright (c) 2022, Keybase'
 const companyName = 'Keybase, Inc.'
 
@@ -53,8 +54,13 @@ const packagerOpts: Options = {
   darwinDarkModeSupport: true,
   dir: desktopPath('./build'),
   download: {
-    mirrorOptions: {
-      mirror: 'https://kbelectron.keybase.pub/electron-download/',
+    checksums: {
+      'electron-v20.0.3-darwin-arm64.zip': 'cc7f3caf673f17b887d803114494cb15babac1574e246780070cda5d79138004',
+      'electron-v20.0.3-darwin-x64.zip': '6fd3ab4152a609e4923af5776ef8d2c3fb1d1842616706ac3acd1e78fc2cb2a9',
+      'electron-v20.0.3-linux-arm64.zip': '4b92fe8a46c937fcea3f689f5498e6978da763653d2e314001800b436a0e1fe7',
+      'electron-v20.0.3-linux-x64.zip': '2fd2bf1bcfc05ea37b31cab23b4e5a3081b5e974ced16d5fdd34b09e6d70b2e8',
+      'electron-v20.0.3-win32-x64.zip': '77b9879f2d9841a41b4daefd15ab9f342dc798066ecf9964ab0979b4173a2463',
+      'hunspell_dictionaries.zip': '211f39a653f75fc29eaf8b7effcb428d67252c232a5344dd1c7a8ed311062987',
     },
   },
   electronVersion: undefined,
@@ -97,17 +103,26 @@ const packagerOpts: Options = {
   ],
 }
 
-function main() {
-  rimraf.sync(desktopPath('dist'))
-  rimraf.sync(desktopPath('build'))
+async function main() {
+  if (TEMP_SKIP_BUILD) {
+    for (let i = 0; i < 10; ++i) {
+      console.log('TEMP_SKIP_BUILD true@!!')
+    }
+  } else {
+    rimraf.sync(desktopPath('dist'))
+    rimraf.sync(desktopPath('build'))
+  }
 
   copySync('Icon.png', 'build/desktop/Icon.png')
   copySync('Icon@2x.png', 'build/desktop/Icon@2x.png')
   copySyncFolder('../images', 'build/images', ['.gif', '.png'])
-  fs.removeSync(desktopPath('build/images/folders'))
-  fs.removeSync(desktopPath('build/images/iconfont'))
-  fs.removeSync(desktopPath('build/images/mock'))
-  fs.removeSync(desktopPath('build/desktop/renderer/fonts'))
+  if (TEMP_SKIP_BUILD) {
+  } else {
+    fs.removeSync(desktopPath('build/images/folders'))
+    fs.removeSync(desktopPath('build/images/iconfont'))
+    fs.removeSync(desktopPath('build/images/mock'))
+    fs.removeSync(desktopPath('build/desktop/renderer/fonts'))
+  }
 
   fs.writeJsonSync(desktopPath('build/package.json'), {
     main: 'desktop/dist/node.bundle.js',
@@ -115,8 +130,8 @@ function main() {
     version: appVersion,
   })
 
-  const icon: string = argv.icon
-  const saltpackIcon: string = argv.saltpackIcon
+  const icon: string = argv.icon ?? ''
+  const saltpackIcon: string = argv.saltpackIcon ?? ''
 
   if (icon) {
     packagerOpts.icon = icon
@@ -141,26 +156,34 @@ function main() {
   }
 
   try {
-    startPack()
+    await startPack()
   } catch (err) {
     console.log('Error startPack: ', err)
     process.exit(1)
   }
 }
 
-function startPack() {
+async function startPack() {
   console.log('Starting webpack build\nInjecting __VERSION__: ', appVersion)
   process.env.APP_VERSION = appVersion
   const webpackConfig = rootConfig(null, {mode: 'production'})
-  webpack(webpackConfig, (err, stats) => {
-    if (err) {
-      console.error(err)
-      process.exit(1)
-    }
+  try {
+    if (TEMP_SKIP_BUILD) {
+    } else {
+      const stats = await new Promise<webpack.Stats | undefined>((resolve, reject) => {
+        webpack(webpackConfig, (err, stats: webpack.Stats | undefined) => {
+          if (err) {
+            reject(err)
+          } else {
+            resolve(stats)
+          }
+        })
+      })
 
-    if (stats?.hasErrors()) {
-      console.error(stats.toJson('errors-only').errors)
-      process.exit(1)
+      if (stats?.hasErrors()) {
+        console.error(stats.toJson('errors-only').errors)
+        process.exit(1)
+      }
     }
 
     copySyncFolder('./dist', 'build/desktop/sourcemaps', ['.map'])
@@ -168,39 +191,34 @@ function startPack() {
     fs.removeSync(desktopPath('build/desktop/dist/fonts'))
 
     rimraf.sync(desktopPath('release'))
-    if (shouldBuildAll) {
-      // build for all platforms
-      const aps = [
+
+    let aps = [[platform, arch]]
+    if (platform === 'darwin') {
+      aps = [
         ['x64', 'darwin'],
         ['arm64', 'darwin'],
-        ['ia32', 'linux'],
-        ['x64', 'linux'],
-        ['x64', 'win32'],
       ]
-      aps.forEach(([arch, plat]) => {
-        pack(plat, arch).then(postPack(plat, arch)).catch(postPackError)
-      })
-    } else {
-      // build both on macos
-      if (platform === 'darwin') {
-        const aps = [
-          ['x64', 'darwin'],
-          ['arm64', 'darwin'],
-        ]
-        aps.forEach(([arch, plat]) => {
-          pack(plat, arch).then(postPack(plat, arch)).catch(postPackError)
-        })
-      } else {
-        pack(platform, arch).then(postPack(platform, arch)).catch(postPackError)
-      }
     }
-  })
+    await Promise.all(
+      aps.map(async ([arch, plat]) => {
+        try {
+          const appPaths = await pack(plat, arch)
+          postPack(appPaths, plat, arch)
+        } catch (err) {
+          console.error(err)
+          process.exit(1)
+        }
+      })
+    )
+  } catch (err) {
+    console.error(err)
+    process.exit(1)
+  }
 }
 
-// eslint-disable-next-line
-function pack(plat: string, arch: string): Promise<any> {
+async function pack(plat: string, arch: string) {
   // there is no darwin ia32 electron
-  if (plat === 'darwin' && arch === 'ia32') return Promise.resolve()
+  if (plat === 'darwin' && arch === 'ia32') return []
 
   let packageOutDir = outDir
   if (packageOutDir === '') packageOutDir = desktopPath(`release/${plat}-${arch}`)
@@ -230,39 +248,34 @@ function pack(plat: string, arch: string): Promise<any> {
   return packager(opts)
 }
 
-const postPackError = err => {
-  console.error(err)
-  process.exit(1)
-}
-
-function postPack(plat, arch) {
-  return appPaths => {
-    if (!appPaths || appPaths.length === 0) {
-      console.log(`${plat}-${arch} finished with no app bundles`)
-      return
-    }
-    const subdir = plat === 'darwin' ? 'Keybase.app/Contents/Resources' : 'resources'
-    const dir = path.join(appPaths[0], subdir, 'app/desktop/dist')
-    const modules = ['node', 'main', 'tracker2', 'menubar', 'unlock-folders', 'pinentry']
-    const files = [
-      ...modules.map(p => p + '.bundle.js'),
-      ...modules.filter(p => p !== 'node').map(p => p + '.html'),
-    ]
-    files.forEach(file => {
-      try {
-        const stats = fs.statSync(path.join(dir, file))
-        if (!stats.isFile() && stats.size > 0) {
-          console.error(`Detected a problem with packaging ${file}: ${stats.isFile()} ${stats.size}`)
-          process.exit(1)
-        }
-      } catch (err) {
-        console.error(`${path.join(dir, file)} doesn't exist`)
-        console.error(err)
+function postPack(appPaths: Array<string> | null, plat: string, arch: string) {
+  if (!appPaths || appPaths.length === 0) {
+    console.log(`${plat}-${arch} finished with no app bundles`)
+    return
+  }
+  const subdir = plat === 'darwin' ? 'Keybase.app/Contents/Resources' : 'resources'
+  const dir = path.join(appPaths[0], subdir, 'app/desktop/dist')
+  const modules = ['node', 'main', 'tracker2', 'menubar', 'unlock-folders', 'pinentry']
+  const files = [
+    ...modules.map(p => p + '.bundle.js'),
+    ...modules.filter(p => p !== 'node').map(p => p + '.html'),
+  ]
+  files.forEach(file => {
+    try {
+      const stats = fs.statSync(path.join(dir, file))
+      if (!stats.isFile() && stats.size > 0) {
+        console.error(`Detected a problem with packaging ${file}: ${stats.isFile()} ${stats.size}`)
         process.exit(1)
       }
-    })
-    console.log(`${plat}-${arch} finished!`)
-  }
+    } catch (err) {
+      console.error(`${path.join(dir, file)} doesn't exist`)
+      console.error(err)
+      process.exit(1)
+    }
+  })
+  console.log(`${plat}-${arch} finished!`)
 }
 
 main()
+  .then(() => {})
+  .catch(() => {})
